refactor:접미사 처리 함수 공용 모듈로 분리
